pyproject.toml
[project]
name = "SICrMLB"
version = "0.1.0"
description = "Default template for PDM package"
authors = [
    {name = "artumont", email = "197291181+artumont@users.noreply.github.com"},
]
dependencies = ["pydantic>=2.12.2", "numpy>=2.3.4", "opencv-python>=4.11.0.86", "av>=16.0.1", "pillow>=12.0.0", "pytest>=8.4.2"]
requires-python = "==3.13.*"
readme = "README.md"
license = {text = "MIT"}


[tool.pdm]
distribution = false
